requirements.txt
automata-lib[visual]==4.0.0